requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2
django-cors-headers==4.7.0
django-haystack==3.3.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
mysqlclient==2.2.7
packaging==24.2
PyJWT==2.9.0
pysolr==3.10.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
